# Use const generics to remove BitTree heap allocations