--- a/.github/workflows/tests.yml
+++ b/.github/workflows/tests.yml
@@ -11,7 +11,7 @@ jobs:
           - stable
           - beta
           - nightly
-          - 1.50.0  # MSRV
+          - 1.57.0  # MSRV
       fail-fast: false
     runs-on: ${{ matrix.os }}
     env: