@chyyran 2022-08-09
https://github.com/gendx/lzma-rs/pull/79/commits/37e09d126b83bbb390763af5a24626c3ab1dde98 represents an alternative API with `BitTree` parameterized on a single generic, which is unfortunately not `NUM_BITS` but `PROBS_ARRAY_LEN`  as we can't use associated constants in struct definitions. This relies on the idea that that `1 << n` has the inverse `(1 << n).trailing_zeroes()`, both of which are const. We can then use `NUM_BITS` as an associated constant.

The signature of `BitTree` here is cleaner, but the downside is that it is now possible to construct invalid instances of `BitTree` since there is no way to check that `PROBS_ARRAY_LEN` is the intended value without passing in `NUM_BITS`, and there is no way to check for this. For this reason I prefer the former implementation, but since `BitTree` is an internal API this really is just a matter of style and clarity.