@chyyran 2022-08-11
https://github.com/gendx/lzma-rs/pull/79/commits/9223b59a1fd13b28bb88aa27a632f7dbc867fd9c introduces a combining of both approaches. While the generic argument is still `1 << NUM_BITS` (removed the helper const function since I thought it was clear enough without it), I realized there can still be some safety had to ensure that as long as `trailing_zeros` was valid for log_2 the input array length would be valid.