@chyyran 2022-09-03
Marking this as ready since #77 was merged.

@gendx you will want to review both 06aee699384477e09aa98e3d029197e927536b31 and 260eefe24b47e07d8cff2ccc8afda36755ce5ca2 which are identical **except for the signature of `BitTree`**, as explained in the overview. 

Given that `generic_const_exprs` doesn't seem to be coming very soon, and that `BitTree` is an internal API that isn't exposed, I'm of the opinion that the perf improvements are worth the hit to readability, but you may want to consider re-running some benchmarks on your end as well.